APEX-24 #resolve took out rhino dependency as it is GPL
